[build] Add minimal support to build Win32 binaries.
This will be needed to generate the UEFI boot loader, which
must be a Win32 PE32+ binary. This CL only modifies a few
necessary configs when current_os is "win", which none of
the existing toolchain() instances in the Fuchsia build
support anyway.

These changes are here to reflect the content of
//zircon/public/gn/config/BUILD.zircon.gn which provides many
configs that are similar to the ones in
//build/config/BUILD.gn

+ Move //build/config/fuchsia:icf to //build/config:icf
  since this will be used by Zircon-specific toolchains
  as well.

Bug: 54161
Testing: This should not change the build.
Change-Id: Iff4f01dc7230836f40c56c928267bf52cb135b39
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401617
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>