[serial] Add option for server to start at end
As we move to persistent serial logs, we need a multiplexer that
starts new connections at the current point in time. However, the
current behavior is necessary for the tight interval botanist needs
to run in today, so make this an option that can be passed into the
server.

Change-Id: I9507956a6be0d375d7d378cc3aa48697a228c8ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409258
Reviewed-by: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>