[serial] Assume socket server starts at end
Botanist used to have to copy through a large amount of data to get
to the current location in the serial log in the persistent serial case.
In other words, when a socket server was started outside of the task, we
had to process a very large amount of data. We can skip this now that
the serial server supports starting at the end of the aux output.

Change-Id: Ibc616decbcfb6b98ab3bcc33f6a16e0adc7ebe9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409261
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>