[camera] Consolidate zircon.device.camera with ddk.protocol.camera
This CL resolves an old issue. Originally, the redefinition of the same
struct in both FIDL and Banjo interfaces was avoided in favor of using a
C wrapper. This solution is imperfect, e.g. in the instance that hlcpp bindings are used, it will need to be manually serialized anyway. The ISP implementation currently works as is.

This CL also helps pave the way to enabling a FrameRateInfo struct which
is necessary for the low light feature.

Test: n/a
Fixed: 37654

Change-Id: I3b4cccb114520b1c1bc244503a9c20eb3d28ad94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405786
Commit-Queue: Nic Zoghb <nzo@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Brijen Raval <braval@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>