[wlan-hw-sim] Use WLAN Policy updates to detect failed association
The legacy policy::connect() command returns after either [connection is
made successfully] or [timeout after 20 seconds]. Since the connection
will never be made successfully in this test, the command always returns
after 20 seconds. The wlancfg module will continue attempting to scan
and join saved networks in the background.

After those 20 seconds, the test checks the status of the interface to
see if the connection was successful. However, the interface may be
either in SCANNING or JOINING state at this time, because of the
previously described auto-join behavior.

This CL modifies the test to use the modern Policy APIs, which gets
rid of this race condition

Bug: 56104
Test: fx test connect-with-failed-association
Multiply: connect-with-failed-association
Change-Id: I762bf51b3ffe9d6f6f1f6e754d152797c4fffbd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407855
Commit-Queue: Marc Khouri <mnck@google.com>
Reviewed-by: Eric Wang <eyw@google.com>
Testability-Review: Eric Wang <eyw@google.com>