[brcmfmac] Handle failed calls to getentropy
Rather than asserting when a call to getentropy() fails, treat it
as a runtime error and allow it to be handled with normal failure
logic.

Entropy is used to generate a random mac address, which
happens in two instances: device initialization (for our initial
mac address when none is available from the bootloader), and for
an active scan.

To accommodate testing, this CL includes a BRCMF_CALL construct
that allows us to override calls to external libraries in tests.

It also includes a fix to the simulator to properly cleanup when
initialization fails (which wasn't a problem in any of our
previous sim tests).

Bug: 51987
Test: Included (bad_entropy_test.cc). Also ran manual smoke tests
      (active scan, passive scan connect, ping, disconnect) to
      verify that basic operation was not affected.
Change-Id: Ieddc636f0b69557bd605c3cf0e0b7fc1d2fd92d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408954
Commit-Queue: Joshua Conner <joshconner@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Testability-Review: Zhiyi Chen <zhiyichen@google.com>