[ping3] Fake time during SequenceStore tests
A SequenceStore#give test was confusing response times with latencies.
This change corrects that by testing the calculation of latencies. To do
this reliably, an executor with fake time is used for testing.

SequenceStore#give returns a latency in the form of a Duration.
Asserting equality of the two latencies doesn't make sense since they
should actually both be the same in the original test. However the test
wasn't failing 99% of the time since the test runner usually takes
longer than 1ns to reach the next call to SequenceStore#give.

Fixed: 55572

Change-Id: I3ca1adc5cbc90007a387fef18f95764f57ad1262
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408756
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>
Commit-Queue: Sam Balana <sbalana@google.com>