[camera] Add warning for unexpected buffer ID.
This change adds a warning and early-return check in cases where the
driver sends a buffer ID to the HAL client that it has previously sent
but that the client has not released. This indicates non-compliant
behavior by the driver.

Bug: 44037
Test: A future change will more formally validate (and test validation
      of) non-compliant driver behavior. See the above-linked bug.
Change-Id: I8d2c24cd4fe1537adc7e67a58424a976d2d0be1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404662
Reviewed-by: Ernest Hua <ernesthua@google.com>
Testability-Review: Matt Sandy <msandy@google.com>
Commit-Queue: Matt Sandy <msandy@google.com>