[scrutiny] Package added to data model.
This change adds the package data to the data model which provides
useful information about which packages contain or use which shared
objects. This is useful for auditing packages as it allows the user to
quickly determine the scope of a vulnerability in a shared library by
cross referencing it with the packages it is in. I plan to extend this
in a future CL to enable parsing of the ZBI to extract the
component_manager manifest.

Change-Id: Idac72335787416c45c60ed6db1f1a85528267e4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409320
Commit-Queue: Benjamin Wright <benwright@google.com>
Testability-Review: Benjamin Wright <benwright@google.com>
Reviewed-by: Benjamin Wright <benwright@google.com>
Reviewed-by: Daniel Hwang <arkay@google.com>