[camera] Report aspect ratio in controller configs.
This change updates the stream image formats reported by the controller
to include the correct pixel aspect ratio. This allows a client to
display the content with the correct ratio, or select a crop region that
restores a unit ratio if desired.

This change also updates camera-gym to scale scenic nodes according to
the aspect ratio so they are displayed at the correct ratio.

Test: Ran camera-gym and observed corrected display aspect ratio.
Change-Id: I713ae326819cc247d02d1e08e1b728a54302f852
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409146
Commit-Queue: Matt Sandy <msandy@google.com>
Reviewed-by: Brijen Raval <braval@google.com>
Testability-Review: Brijen Raval <braval@google.com>