[scrutiny] Package search.
This adds a simple web view to allow you to search through all available
packages in the system for particular files. This is nice if you want to
search for a vulnerable library and see all the packages that contain
it.

Change-Id: Ia7d4a748357f4f3262c7b6e6fa271de1db053072
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409701
Reviewed-by: Daniel Hwang <arkay@google.com>
Reviewed-by: Benjamin Wright <benwright@google.com>
Testability-Review: Benjamin Wright <benwright@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>