[shutdown-shim] return mexec call when asked to terminate
The mexec shutdown operation is different than the others, in that
driver_manager does not change the system power state during component
shutdown. It merely shuts down drivers and exits, after which the client
which requested the mexec will perform the mexec syscall.

When driving shutdown manually, the shim should notify clients when
mexec work is complete and the client may make the syscall. The shim
does this by initiating a component shutdown, and finishing the mexec
call when the framework asks the shim to exit, as this means that
driver_manager shutdown is complete.

Change-Id: I7b7a58507bf0378e4d51e60bb4024c095c5aed1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407960
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Ruchira Ravoori <ravoorir@google.com>
Reviewed-by: Patrick Shickel <pshickel@google.com>
Testability-Review: Patrick Shickel <pshickel@google.com>