[brcmfmac] Handle failed calls to getentropy
This reverts commit bccd19966fd5003fc9873b6bfa7e4e9c8a1cabf7 (and
redundant revert e27175e3fe25d6d1a5749201b8ca4a3991043943), which
were reverts of 14eafa3690bc4038d189a9d16aca72828b06e2b8, which is
this commit with the following changes:
* Properly release BCDC resources on simulated bus initialization
  failure.

Rather than asserting when a call to getentropy() fails, treat it
as a runtime error and allow it to be handled with normal failure
logic.

Entropy is used to generate a random mac address, which
happens in two instances: device initialization (for our initial
mac address when none is available from the bootloader), and for
an active scan.

To accommodate testing, this CL includes a BRCMF_CALL construct
that allows us to override calls to external libraries in tests.

It also includes a fix to the simulator to properly cleanup when
initialization fails (which wasn't a problem in any of our
previous sim tests).

Bug: 51987, 56856
Test: Included (bad_entropy_test.cc). Also ran manual smoke tests
      (active scan, passive scan connect, ping, disconnect) to
      verify that basic operation was not affected.
Change-Id: Ia1735e449089e298b008ba49b9ea1d6f5ee08abb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409934
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Testability-Review: Zhiyi Chen <zhiyichen@google.com>
Commit-Queue: Joshua Conner <joshconner@google.com>