[kernel][pc] Fix spinlock contention in UART IRQ handler.
Fix over synchronization between TX and IRQ paths that causes the
IRQ handler to spin for long periods of time on a spinlock when
there is a lot of TX activity. Because IRQs are handled on CPU 0
this can cause tasks that are assigned to CPU 0 to be significantly
delayed by the time "stolen" by the spinning IRQ handler.

Synchronization within the IRQ handler is unnecessary because the
TX, RX, and IRQ paths do not perform any read-modify-write ops on
UART registers and the TX and RX functions are independent. Also,
the cbuf and event objects shared between the IRQ handler and TX
and RX operations have internal spinlocks to protect invariants.

Fixed: 55085
Test: Observed traces of continuous TX activity, written from both
thread and IRQ context. Spinning is observed before this change but
not after.

Change-Id: I7b59deae58c427c99143666f00747446547831bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409519
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>