[servicefs] Allow passthrough of flags intended for remote dirs
ServiceFS checks flags on every Open request irrespective of
whether the path goes through a remote directory.

For example, when attempting to create a file in a remote
directory (such as a filesystem), ServiceFS will throw an
error even though the CREATE flag was intended for the
remote directory and not for ServiceFS itself.

To solve this, now ServiceFS first descends through the path and
then checks the flags only if it is opening a local directory entry.

Test: `fx test fuchsia-component-tests`
Change-Id: Iddcb73e841b9666a0bd7e2643153bb00493a0dba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409928
Reviewed-by: Fady Samuel <fsamuel@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>