[i18n] Allows lenient localization.
Originally, the l10n JSON files generator errored out if it noticed
an original string that has no translated equivalent.  The intention
there was to point out that translations were missing.

However, this does not work in daily development, where we must have a
temporary discrepancy between the original and translated files, while
we wait for translations to be delivered.

This change introduces a flag `--substitute-original` which substitutes
"UNTRANSLATED(foo)" for message "foo" if a translation is not available.
The idea is to offer a clear way to know that there are faulty strings.
In this case, the expectation is that we could find them in QA.

It is always possible to re-run the tool in a checking mode without this
flag.  There could be other modes, e.g. omit missing string, but we will
not add those ahead of time.

Fixed: 56958
Change-Id: I5137ad1981167f1611ad81fdf5c1dcfcf85a7784
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410100
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Testability-Review: Filip Filmar <fmil@google.com>