[fidl][llcpp] fix ResultOf
The buffer in ResultOf is only valid if the call/write was successful
and, in case of call, if the decoding was successful.

That means that the buffer is valid only if |ok()|.

We have to release handles only if the buffer is ok.

Change-Id: Ib578f619f6475d96ab1ba6eb26a1073d90f37a7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410124
Reviewed-by: Yifei Teng <yifeit@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>
Commit-Queue: Vincent Belliard <vbelliard@google.com>