[GN SDK] Handling gsutil missing in pave and serve
This changes the fuchsia-common functions for running gsutil
and cipd to return a non-zero code when the commands cannot be
found instead of exiting.

This is in response to a developer report of fpave.sh just returning
without an error or success.

Change-Id: Ic51b196dfd55c030ad51f60aaa5950d8ba406855
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410194
Reviewed-by: Wayne Piekarski <waynepie@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>