[fidl][go] Allow unused reflect import in conformance tests
When certain GIDL tests are commented out, this import is unused which
means that you need to go into the go geneartor and temporarily comment
out the import. In order to avoid this, always have a dependency on the
reflect package.

Change-Id: I3b0c0254ee050ee0c576eab029a7ded16b2082af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410048
Reviewed-by: Mitchell Kember <mkember@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>