[time][recovery] Group timesync packages needed for recovery
On a normal Fuchsia build, time synchronization packages are
included on the build via the core product definition. Recovery
images do not depend on core and therefore must include the packages
explicitly.
This change provides a gn group used by recovery image definitions so
that the recovery build rules do not need to be updated when the set
of time sync components changes, which will occur in upcoming changes.

Change-Id: Id97eca1163fb400fbd7e3cefcad25142ba615ef4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407281
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>