[rust][fasync] Prepare for the removal of fasync::spawn*
TLDR: fasync::spawn functions are going away in preference of the new
Task::spawn API's. This CL prepares the way by replacing usage of the old
API's with usage of the new one.

We're deprecating and replacing the old spawn functions (spawn, spawn_local,
spawn_blocking) and replacing them with an API centered around Task objects.

Tasks own the execution of the underlying future, and if dropped will cancel
that execution. To regain the old, unowned execution that spawn API's offered
a detach method is offered.

This change series sweeps through the codebase and replaces spawn() with
Task::spawn().detach() and spawn_local() with Task::local().detach() in
preparation for the removal of the original API's.

It's hoped that code maintainers will then take the opportunity (when it makes
sense for them to do so) to move to a model where Tasks are not detached but
owned for as long as they need to be running.

Change-Id: Ie0297a1b086dbf9d4d42f6b87f149b0956956621
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410215
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Patrick Shickel <pshickel@google.com>
Testability-Review: Derek Gonyeo <dgonyeo@google.com>
Testability-Review: Craig Tiller <ctiller@google.com>
Commit-Queue: Craig Tiller <ctiller@google.com>