[triage] Short circuit boolean operation evaluation
This CL ensures that we don't attempt to evaluate metrics in boolean
operation after one metric would determine the result (a single False in
an And or a single True in an Or).

This can be used to ensure that missing metrics are suppressed if an
earlier defined metric in the boolean operation sets the result.

This also adds unit tests for various metric operations, including the
new behavior.

Change-Id: I4b78a3add3956cba876d5df1c18d1eacd7fb7fa4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402133
Commit-Queue: Christopher Johnson <crjohns@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Chris Phoenix <cphoenix@google.com>
Testability-Review: Chris Phoenix <cphoenix@google.com>