[kernel][iommu] Improve comments for whole_segment field
IOMMU descriptors have a list of scopes (devices) associated with
each segment. whole_segment describes whether the list of scopes
is the list of devices managed by the IOMMU or the list of devices
not managed by the IOMMU.

Improve the comments to make that clearer and to align with the
respectful code guidelines.

Change-Id: I7cac8b3c76cd11e4d5b7e43a1a38d1644e192456
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410197
Reviewed-by: Tess Eisenberger <teisenbe@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Testability-Review: Tess Eisenberger <teisenbe@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>