[ffx] daemon start starts a daemon
The pid must be waited on unix hosts, otherwise the child process
becomes a zombie. This implementation will perform a daemonize
downstream of the fork, so it will succeed if the second fork works,
but will wait for the first fork to complete, avoiding the zombie
condition. This also prevents the hosting session leader from
terminating the daemon at some later point. The allowed deprecation
is for libc::daemonize that is deprecated as of macOS 10.5. We'll
need to bring in a 3p lib or write more ffi code to replace this
eventually.

Change-Id: I695bac3cd9f772aba007a2e081221903f65ad13c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409054
Commit-Queue: James Tucker <raggi@google.com>
Testability-Review: James Tucker <raggi@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>