[debugger] Remove ServiceDirectory from DebugAgent
The DebugAgent should not be using the system FIDL interface for
ServiceDirectory but functionality should be instead exposed by the
SystemInterface object.

Move handler querying out of DebuggedProcess and into ProcessHandle.
This is the last significant system call in the DebuggedProcess which
means it can be completely mocked.

Change-Id: I1a96602eaf005c704ec69f668c0124edff292a36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410262
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>