[third_party][fatfs] Use 0 for root dir in .. entries.
Fat expects all '..' entries which reference the root dir
to have a cluster number of zero, even though on Fat32 this is not
necessarily the case. Fix this by checking if the parent cluster is the
root cluster when creating a directory, and setting cluster to zero if
so.

Bug: 53796
Change-Id: Ie0e2ff8ba480fb215e1095b6471a1e79a5cdf1d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410160
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>