[amlogic_decoder] decode flake repro test
Plumb a test hooks struct.  On astro this demonstrates occasional
incorrect decode of the last frame in the sequence, and when the
decode is incorrect, it's always the same slightly bad pixels on
astro.  On sherlock it's not always the same bad pixels, and the
repro rate is lower.

BUG: 13483
Change-Id: I45ec9b458103124555633b9853f46a3578b30436
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410220
Reviewed-by: John Bauman <jbauman@google.com>
Testability-Review: John Bauman <jbauman@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>