[dart_tool] conditionally generate a library
The dart tool does not need to generate a library unless it
defines sources. This will check if there are sources and
generate a library if needed. This allows us to depend on
multiple dart_tools that exist in the same library from
a single target.

Change-Id: Ifac68dff38009310eaa5429d6023f3314f8b5f4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410896
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>