[camera] Temporarily Disable DeviceEnumerationTest for Camera Devices
This CL removes devices from being run in DeviceEnumerationTest. This is
a temporary change to enable the soft transition of camera drivers to
ddk.protocol.camera.sensor.

WARNING: This is a breaking change for camera functionality and will be undone in a follow-up CL.

Test: n/a

Change-Id: I53dc3b95583a53fe0ac22e2c43e035ff98529221
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410781
Commit-Queue: Nic Zoghb <nzo@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>