[backtrace] Simplify backtrace request output.
The backtrace request library triggers an exception that is specially
handled to resume. But most people don't know this and the output looks
almost identical to a crash, is reported as an exception and a "sw
breakpoint". This results in confusion, especially when used as a
watchdog (Scenic does this).

This removes the output from the crash service that it's received an
exception in the bcase of a backtrace request.

In the inspect output, this clarifies that the result is a backtrace
request and omits the CPU registers and stack binary dump. Normally in a
backtrace request, only the actual stack frames are needed and this
output makes it more difficult to follow and look more like a crash.

Change-Id: I4337a5103a38cde9b35c2a13302e07fb9cf5027c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410821
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Testability-Review: Francois Rousseau <frousseau@google.com>