[block-verity] Implement CloseAndGenerateSeal
Upon request, the driver will now generate integrity data, write it out
to disk, update the superblock with the device configuration and root
integrity hash, and return the hash of the superblock.

Added tests that verify the contents of the generated integrity data,
superblock, and seal against the values expected given the design
specification.

Added additional tests that verify that on any I/O failure, the
computation is marked as failed and completed and propagates the error
in question.

Thanks to Allison Pearce for tracking down why the integration tests
were failing in work-in-progress versions of this patchset.  She
verified that the driver implementation was correct, determined that the
test failures were due to misuse of FIDL (stack use-after-free) in
block-verity-test.cc, and supplied a correct implementation.

Fixed: 55710
Change-Id: Iba9ab61f0583f7418a9764ae7e723c4d5dbd5895
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409817
Testability-Review: Drew Fisher <zarvox@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>