[ffx] Reland, daemon start starts a daemon
The pid must be waited on unix hosts, otherwise the child process
becomes a zombie. This implementation will perform a daemonize
downstream of the fork, so it will succeed if the second fork works,
but will wait for the first fork to complete, avoiding the zombie
condition. This also prevents the hosting session leader from
terminating the daemon at some later point. The allowed deprecation
is for libc::daemonize that is deprecated as of macOS 10.5. We'll
need to bring in a 3p lib or write more ffi code to replace this
eventually.

The daemon does not chdir("/"), as it does not yet have context
information passed from the frontend sufficient to otherwise discover
the correct environment and configuration.

Bug: 57054
Change-Id: I586743b8bf0e14a2c99832161c3bcd046a830756
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410452
Commit-Queue: James Tucker <raggi@google.com>
Testability-Review: James Tucker <raggi@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>