[blobfs] Pager watchdog
Creates a new watchdog thread which logs any page faults that take
longer than a configured deadline (currently 60s) to satisfy.

When the deadline is exceeded, the watchdog takes a backtrace of all
blobfs threads, and logs the error. In the future we can extend this in
other ways if desired. For now, the watchdog never affects the behavior
of the pager thread, or of page fault handling in general.

Since the blobfs pager being stuck can effectively stall the entire
system, this is an important state to watch for, and will aid if any
bugs result in pager hangs (either bugs in blobfs itself, or cycles
where a process that blobfs' pager thread depends on takes a fault).

Tested: Manually (set small deadline to verify logging, set real
        deadline to verify not-logging); new unit tests.
Bug: 56818
Bug: 56934 # Fixed
Change-Id: I4f694f0cf53fd1d32b2330b30283e7870d1ce614
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410050
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>