[camera] Relax asserts on InterfacePtr::Bind failures.
This change relaxes existing asserts on the return value of InterfacePtr
::Bind calls. Specifically, it allows for ZX_ERR_CANCELED, which can
occur if the default dispatcher for a thread is owned by a loop that is
being shut down. This can occur if an object is destroyed before it is
fully initialized.

Test: Reproduced by rapidly changing the configuration in camera-gym.
      Verified that with fix applied, assert no longer fires.
      Added DeviceTest.BindFailureOk test.
Change-Id: Ic522f4a84ca4584d64a8b4be3cf5e0ad5f48fb2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410288
Testability-Review: Matt Sandy <msandy@google.com>
Commit-Queue: Matt Sandy <msandy@google.com>
Reviewed-by: Ernest Hua <ernesthua@google.com>