[camera] Fix binding leak on pipeline creation failure.
This change fixes a bug in the controller where a pipeline that fails
part-way through creation left the ISP bound to a non-existent node.
This can occur if sysmem is under pressure and an inter-node collection
fails allocation. The change adds a shutdown auto-call, and a new check
on shutdown to only propagate the shutdown call if a child node exists.

Test: Reproduced binding leak by shortening camera-gym cycle time.
      Verified fixed by rapidly cycling configs 100 times, then cycling
        as normal.
Change-Id: Iaa2d5b5c8a1c23e3ca606234a5232bcaf4da4f73
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410264
Commit-Queue: Matt Sandy <msandy@google.com>
Testability-Review: Matt Sandy <msandy@google.com>
Reviewed-by: Brijen Raval <braval@google.com>