[testsharder][testrunner] Record results from all test tries.
Testrunner currently allows retries but only records the result of the
last attempt. This hides flakes, so we should report all test runs and
let the recipes deal with how to handle flakes vs failures.

Testsharder is also updated to use a run_algorithm instead of
max_attempts to determine how to use the `runs` field when running a
test. For example, the run_algorithm will either tell testrunner to keep
running the test `runs` amount of times or to attempt the test up to
`runs` amount of times until it gets a success.

Bug: 56702
Change-Id: Iad10c3fc779828abdf5579e29b162ae76d8b8048
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409059
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>