[fatfs] Fix directory rename issue.
When renaming directories, the ".." entry needs to be updated.

This fixes:

  PersistRenameLoopTest.MultipleRenamesCorrectAfterRemount/
     FatfsRenameDirectory100TimesThrough15Directories

Whilst addressing this, I noticed that the implementation writes
long-file-name entries for the '.' and '..' entries, which might run
afoul of some fsck programs (although not ours), so I fixed that.

Bug: 53796
Change-Id: Icb70f5875a8644916a6687080f5c13e8430a8dc8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411395
Reviewed-by: Simon Shields <simonshields@google.com>
Testability-Review: Simon Shields <simonshields@google.com>
Commit-Queue: Chris Suter <csuter@google.com>