[debugger] Improve suspending.
Clarifies the two types of suspensions. Previously there were several
different ways to suspend and resume a thread, and there was also a flag
to indicate whether the client thought that the thread was suspended.

This change separates "client" suspensions that the zxdb frontend
considers its domain, from "internal" suspensions which happen for
things like stepping over breakpoints. Client suspensions have a simple
on/off state, while internal suspensions use suspension tokens.

Creates a SuspendHandle wrapper for suspension tokens so that these can
be used in mocks.

Removes the existing "ref counted suspend" machinery which emulates part
of this in a complex way. Delete the unit tests for this.

Moves waiting for suspension to the ThreadHandle so this can be mocked.

Change-Id: I27848a7837194ffe2cfd21781a9357fec5e690af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411275
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>