[triage][codelab] Update monikers to match real ones
Current monikers in the bugreport and in the codelab do not match real
ones. A previous CL (https://fxrev.dev/407292) updated the solution
to use real monikers. This CL makes sure that the codelab uses them as
well.

An advantage of doing this instead of keeping a fake example is that
the codelab resulting triage file can be used in real-world bug reports.

Change-Id: I622893ff7fbcd5bb3551e0ebc11b88afd363f916
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411560
Reviewed-by: Christopher Johnson <crjohns@google.com>
Testability-Review: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>