[docs] Replacing "SDK" where we mean "IDK"
To make the separation of concerns between the "Core" SDK and the
"Frontend" SDKs, The "Core" SDK is now referred to as
Integrator Development Kit (IDK).

The IDK is the development environment independent artifacts that
define the public Fuchsia platform. This includes system libraries,
headers, FIDL files, and Fuchsia specific host tools needed to target
Fuchsia as a developer.

The IDK is the foundation of SDKs created by development environment
integrators which augment and transform the IDK to be usable in a
specific development environment.

This is being done now as we continue to refine the processes
and stuctures around delivering an end-developer ready SDK.

Bug: 51916

Change-Id: Ib6fe47d837e123d5631c95972fe0424477bb2122
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409263
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Chris Osborn <osbornc@google.com>
Reviewed-by: Jared Bhatti <jaredb@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Claudio Cherubino <ccherubino@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>