[cobalt] Convert the prober into an e2e test.
In this change, a new package is created to run the cobalt_testapp in prober
mode without using run-test-component. Since the prober requires network
access, it is expected to not be run as a test (the test component
framework recently stopped allowing tests to access the network services).
This new non-test package is then run by a new end-to-end test for the
prober. E2E tests are allowed to access network services.

Subsequent changes will update the infra to use the new end-to-end test in
place of the current test package for the cobalt prober CI, and then remove
the old prober package from fuchsia.

Bug: 57076
Change-Id: I51a3960e7cdfffcfa7e0ac1df3ad4e37c52f5e25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410943
Reviewed-by: Zach Bush <zmbush@google.com>
Reviewed-by: Isma Cortes <isma@google.com>
Testability-Review: Zach Bush <zmbush@google.com>
Commit-Queue: Cameron Dale <camrdale@google.com>