[GN SDK] Improving fserve-remote
This clears the default IP address and sets the default device name on
the laptop running fserve-remote. Then on the remote desktop, the
default IP address is set, and the device name cleared.

This eliminates any stale default values which could conflict with the
remote serving.

This also adds the sleep invocation to keep the tunnel open vs. closing
quickly.

It also adds "-i" to the shebang for fserve.sh so it uses an interactive
shell, meaning the path and other environment variables are setup
correctly. This fixes the issues with not finding gsutil.

BUG: b/162249694
Change-Id: I092bc23391dd8b57fabd43b8517706b2237a1f8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410994
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Wayne Piekarski <waynepie@google.com>
Reviewed-by: Jonathan Koren <jdkoren@google.com>
Testability-Review: Jonathan Koren <jdkoren@google.com>