[drm] Add KeySystem.DestroyDataStoreAndConfirm
When a data store is added to a KeySystem, it includes a
fuchsia.io.Directory that is used for storing data. When this data
store is destroyed, the KeySystem should clear the contents of this
directory and release all references to it. By sending a reply after
this is complete, the client can ensure that it has been properly
cleared.

Bug: 13619
Change-Id: I0255c3b36f1bb95e7433a54695b63471a2492d0a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412045
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Reviewed-by: Wez . <wez@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>
API-Review: Wez . <wez@google.com>