[tefmocheck] Add a check for QEMU segfaults
This is one of the more common unattributed flakes I've noticed. For
example: https://ci.chromium.org/b/8873278597020207616

I also reworked the way botanist logs these errors so it doesn't seem
like they're coming from testrunner. Specifically, we only log "command
[testrunner] failed: <err>" if the error that occurred is a subprocess
ExitError.

In the case of this failure mode, the log previously looked like:

  botanist ERROR: command [testrunner ...] failed: QEMU invocation error: signal: segmentation fault

but now it will look like:

  botanist ERROR: QEMU invocation error: signal: segmentation fault

This is better because the error is unrelated to the command being run,
so the old log was misleading. It also makes it easier to match this
string in tefmocheck, since there's no command string in the middle of
the log.

Bug: 57463
Change-Id: I65281412e63794d6a3eb17861cda1de611722b46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412499
Reviewed-by: Gary Miguel <garymm@google.com>
Testability-Review: Gary Miguel <garymm@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>