[Docs] RFC: Units of bytes
Why is this a priority for current work: This issue was discovered
because memory is at a premium. While debugging, it was found that
different tools gave values for byte counts in different notation.
These discrepancies are large enough to cause concern and confusion.
Being constant in our measurement notation is expected to streamline
this work.

Change-Id: I501cc99e119f079fa5122bde895f370438e5753a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397239
Commit-Queue: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Nathaniel Manista <nathaniel@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Jared Bhatti <jaredb@google.com>
Reviewed-by: Kristin Gray <heygray@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Alex Bogart <abbogart@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>