[build] fix kasan-sancov variant.
This CL modifies the kasan-sancov variant so that it can boot.

The CL removes the sancov config from the kasan runtime, and the asan
config from the sancov runtime.

Note that we need to remove the entire asan config from the sancov
runtime, as annotating the sancov functions with `NO_ASAN` is not
enough: if the sancov functions are annotated with `NO_ASAN`, calls to
`ktl::atomic` operations within the sancov runtime won't be inlined, and
they end up being seperate asan-instrumented function calls. During
early boot, this would cause a crash, as the memory accesses that happen
before the kernel is relocated cannot be checked by the asan runtime.

Removing the sancov instrumentation from the asan runtime is fine: we
don't need code coverage information of it, and we don't want to make
the code slower.

To test this, I used:

```
$ fx set core.x64 \
  --with //bundles:tools \
  --variant kasan-sancov
```

And it boots and runs all kernel unit tests. (`k ut all`)

Change-Id: Ied1471d80d2f6ed2ca3a3754b872f2008a9bbceb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412130
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>