[cmc] cmc accepts name or path for routing protocol and directory
As an intermediate step, allow cmc to accept use/offer/expose
protocol/directory to reference a capability name or a path.

This CL does not include cm_fidl or routing changes.

Note: This change adds a noticeable amount of complexity to handle both
names and paths. We expect this to be short lived, as the path support
will be deleted once we complete the transition.

Bug: 56604
Change-Id: I283bedc10878c72fad404471c4ee4d7ebfe726e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410949
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Testability-Review: Derek Gonyeo <dgonyeo@google.com>