[zxdb] Handle new tombstone value
With https://reviews.llvm.org/D81784, -1 is a reserved value to indicate
a removed function. zxdb initially tried to use this as the starting
address for an address range.

This patch applies a fix similar to breakpad where we ignore entries
with a value of -1.

TEST=Checked locally to assert we don't see this error with the new
clang and llvm prebuilts.
Bug: 56893

Change-Id: I6650639acd86a0263d861ffcd4812d60a98a96f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412694
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>
Testability-Review: Brett Wilson <brettw@google.com>